tests/python/contract/test_constitutional_contracts.py
#!/usr/bin/env python3
"""
Constitutional Contract Tests - RED Phase (Must Fail Initially)
Tests the API contracts and constitutional compliance interfaces.
These tests MUST FAIL until constitutional enforcement is implemented.
"""

import pytest
import json
from pathlib import Path
from typing import Dict, Any, List


class TestConstitutionalContracts:
    """Contract tests for constitutional compliance enforcement."""

    def setup_method(self):
        """Setup constitutional contract testing environment."""
        self.constitutional_agent_path = Path(".claude/agents/development/constitutional-enforcement-agent.md")
        self.project_constitution_path = Path(".claude/context/project-constitution.md")

    def test_constitutional_enforcement_agent_contract(self):
        """Test Constitutional Enforcement Agent API contract - MUST FAIL initially."""
        # RED PHASE: This test must fail until agent is implemented
        try:
            from src.agents.constitutional_enforcement import ConstitutionalEnforcementAgent
            agent = ConstitutionalEnforcementAgent()

            # Contract: Agent must have supreme authority
            assert hasattr(agent, 'supreme_authority'), "Agent must have supreme authority"
            assert agent.authority_level == "SUPREME", "Authority level must be SUPREME"

            # Contract: Agent must enforce non-negotiable principles
            principles = agent.get_constitutional_principles()
            required_principles = [
                "LIBRARY_FIRST_ARCHITECTURE",
                "TDD_REQUIRED_NON_NEGOTIABLE",
                "TEST_ORDER_HIERARCHY",
                "REAL_DEPENDENCIES_INTEGRATION",
                "MODULE_COMMUNICATION_VIA_EVENTS",
                "OPAQUE_TOKENS_ONLY",
                "GDPR_COMPLIANCE",
                "OBSERVABILITY_REQUIRED"
            ]

            for principle in required_principles:
                assert principle in principles, f"Constitutional principle {principle} must be enforced"

            # Contract: Agent must have violation detection
            assert hasattr(agent, 'detect_violations'), "Agent must detect constitutional violations"
            assert hasattr(agent, 'enforce_compliance'), "Agent must enforce compliance"
            assert hasattr(agent, 'block_non_compliant_code'), "Agent must block non-compliant code"

        except ImportError:
            # Expected failure in RED phase
            pytest.fail("Constitutional Enforcement Agent not implemented - RED phase expected")

    def test_tdd_compliance_contract(self):
        """Test TDD Compliance enforcement contract - MUST FAIL initially."""
        # RED PHASE: This test must fail until TDD enforcement is implemented
        try:
            from src.agents.tdd_compliance import TDDComplianceAgent
            agent = TDDComplianceAgent()

            # Contract: Must enforce RED-GREEN-Refactor cycle
            assert hasattr(agent, 'validate_red_phase'), "Must validate RED phase"
            assert hasattr(agent, 'validate_green_phase'), "Must validate GREEN phase"
            assert hasattr(agent, 'validate_refactor_phase'), "Must validate REFACTOR phase"

            # Contract: Must enforce test hierarchy
            hierarchy = agent.get_test_hierarchy()
            expected_order = ["CONTRACT", "INTEGRATION", "E2E", "UNIT"]
            assert hierarchy == expected_order, f"Test hierarchy must be {expected_order}"

            # Contract: Must block implementation without tests
            assert hasattr(agent, 'block_implementation_without_tests'), "Must block implementation without tests"

        except ImportError:
            # Expected failure in RED phase
            pytest.fail("TDD Compliance Agent not implemented - RED phase expected")

    def test_module_boundary_contract(self):
        """Test module boundary enforcement contract - MUST FAIL initially."""
        # RED PHASE: This test must fail until module boundaries are implemented
        try:
            from src.architecture.module_boundaries import ModuleBoundaryEnforcer
            enforcer = ModuleBoundaryEnforcer()

            # Contract: Must define module boundaries
            modules = enforcer.get_defined_modules()
            required_modules = ["auth", "payment", "user", "subscription", "audit", "shared"]

            for module in required_modules:
                assert module in modules, f"Module {module} must be defined"

            # Contract: Must prevent cross-module dependencies
            assert hasattr(enforcer, 'validate_no_cross_module_deps'), "Must validate no cross-module dependencies"
            assert hasattr(enforcer, 'enforce_event_communication'), "Must enforce event communication"

        except ImportError:
            # Expected failure in RED phase
            pytest.fail("Module Boundary Enforcer not implemented - RED phase expected")

    def test_multi_agent_coordination_contract(self):
        """Test multi-agent coordination contract - MUST FAIL initially."""
        # RED PHASE: This test must fail until coordination is implemented
        try:
            from src.agents.task_coordination import TaskCoordinationAgent
            agent = TaskCoordinationAgent()

            # Contract: Must coordinate multiple agents
            assert hasattr(agent, 'coordinate_agents'), "Must coordinate multiple agents"
            assert hasattr(agent, 'orchestrate_workflow'), "Must orchestrate workflows"

            # Contract: Must support coordination patterns
            patterns = agent.get_coordination_patterns()
            required_patterns = ["SEQUENTIAL", "PARALLEL", "ORCHESTRATED"]

            for pattern in required_patterns:
                assert pattern in patterns, f"Coordination pattern {pattern} must be supported"

        except ImportError:
            # Expected failure in RED phase
            pytest.fail("Task Coordination Agent not implemented - RED phase expected")

    def test_security_compliance_contract(self):
        """Test security compliance enforcement contract - MUST FAIL initially."""
        # RED PHASE: This test must fail until security compliance is implemented
        try:
            from src.security.compliance_enforcer import SecurityComplianceEnforcer
            enforcer = SecurityComplianceEnforcer()

            # Contract: Must enforce OWASP Top 10
            assert hasattr(enforcer, 'validate_owasp_compliance'), "Must validate OWASP compliance"

            # Contract: Must enforce PCI DSS
            assert hasattr(enforcer, 'validate_pci_dss_compliance'), "Must validate PCI DSS compliance"

            # Contract: Must enforce opaque tokens
            assert hasattr(enforcer, 'enforce_opaque_tokens'), "Must enforce opaque tokens"
            assert hasattr(enforcer, 'block_jwt_implementations'), "Must block JWT implementations"

        except ImportError:
            # Expected failure in RED phase
            pytest.fail("Security Compliance Enforcer not implemented - RED phase expected")

    def test_event_driven_architecture_contract(self):
        """Test event-driven architecture contract - MUST FAIL initially."""
        # RED PHASE: This test must fail until event architecture is implemented
        try:
            from src.events.event_coordinator import EventCoordinator
            coordinator = EventCoordinator()

            # Contract: Must support event publishing
            assert hasattr(coordinator, 'publish_event'), "Must support event publishing"
            assert hasattr(coordinator, 'subscribe_to_event'), "Must support event subscription"

            # Contract: Must prevent direct service calls
            assert hasattr(coordinator, 'block_direct_service_calls'), "Must block direct service calls"

        except ImportError:
            # Expected failure in RED phase
            pytest.fail("Event Coordinator not implemented - RED phase expected")


class TestConstitutionalAPIContracts:
    """API contract tests for constitutional enforcement endpoints."""

    def test_constitutional_validation_api_contract(self, api_server):
        """Test constitutional validation API contract - MUST FAIL initially."""
        # RED PHASE: API endpoints must not exist yet
        import requests

        # Contract: Constitutional validation endpoint
        response = requests.post("http://localhost:8080/api/constitutional/validate",
                               json={"code": "sample code", "module": "payment"})

        # Expected contract response
        assert response.status_code == 200
        result = response.json()

        required_fields = ["compliant", "violations", "enforcement_actions", "authority"]
        for field in required_fields:
            assert field in result, f"Response must contain {field}"

    def test_tdd_enforcement_api_contract(self, api_server):
        """Test TDD enforcement API contract - MUST FAIL initially."""
        import requests

        # Contract: TDD validation endpoint
        response = requests.post("http://localhost:8080/api/tdd/validate",
                               json={"test_files": [], "implementation_files": []})

        assert response.status_code == 200
        result = response.json()

        required_fields = ["tdd_compliant", "phase", "violations", "required_actions"]
        for field in required_fields:
            assert field in result, f"TDD response must contain {field}"


if __name__ == "__main__":
    pytest.main([__file__, "-v"])