requirements-dev.txt
# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-json-report==1.5.0
flask==3.0.0
pyyaml==6.0.1

# Security tools
bandit==1.7.5
safety==2.2.0
flake8==6.1.0

# Code formatting
black==23.10.1
isort==5.12.0

# Type checking
mypy==1.5.1
types-PyYAML==6.0.12.12
types-Flask==1.1.6

# Dependencies
packaging>=21.0